get house info querry

+++ sql/deleteSavedhouse.sql
@@ -1,2 +1,2 @@
 -- delete from favorites table
-DELETE FROM `User_favorites` WHERE `House_id` = '2001'
+DELETE FROM `User_favorites` WHERE `House_id` = '2001' AND `User_id` = '222'
